@@ -1,15 +1,16 @@
 package br.com.jacksonroberio.jackar;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import android.annotation.SuppressLint;
 import android.os.Bundle;
 import android.view.Window;
 import android.view.WindowManager;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
+@SuppressLint("SetJavaScriptEnabled")
 public class JacKar extends AppCompatActivity {
-
-    private WebView jogo;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -18,13 +19,12 @@ public class JacKar extends AppCompatActivity {
         requestWindowFeature(Window.FEATURE_NO_TITLE);
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN, WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_jackar);
-
-        jogo = (WebView) findViewById(R.id.webview_jogo);
+        WebView jogo = (WebView) findViewById(R.id.webview_jogo);
 
         jogo.getSettings().setJavaScriptEnabled(true);
 
         jogo.setWebViewClient(new WebViewClient());
 
-        jogo.loadUrl("https://jacksonroberio.com.br/jackar/");
+        jogo.loadUrl("https://jacksonroberio.com.br/jackar/mobile/android/");
     }
 }
